+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/jdk/SecurityProvidersSupport.java
@@ -28,24 +28,28 @@ package com.oracle.svm.core.jdk;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.security.Provider;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 
+import org.graalvm.collections.EconomicMap;
+import org.graalvm.collections.EconomicSet;
 import org.graalvm.nativeimage.ImageSingletons;
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
 
+import com.oracle.svm.core.util.ImageHeapMap;
 import com.oracle.svm.core.util.VMError;
 
 import jdk.graal.compiler.api.replacements.Fold;
+import sun.security.util.Debug;
 
 /**
- * The class that holds various build-time and run-time structures necessary for security providers.
+ * The class that holds various build-time and run-time structures necessary for security providers,
+ * but only in case they are initialized at run time (see the <a href=
+ * "../../../../../../../../../../../../docs/reference-manual/native-image/JCASecurityServices.md">
+ * JCA Security Services documentation</a> for details).
  */
 public final class SecurityProvidersSupport {
     /**
@@ -57,10 +61,10 @@ public final class SecurityProvidersSupport {
      * SecurityServicesFeature#registerServiceReachabilityHandlers).
      */
     @Platforms(Platform.HOSTED_ONLY.class)//
-    private final Set<String> markedAsNotLoaded = Collections.synchronizedSet(new HashSet<>());
+    private final Set<String> markedAsNotLoaded = ConcurrentHashMap.newKeySet();
 
     /** Set of fully qualified provider names, required for runtime resource access. */
-    private final Set<String> userRequestedSecurityProviders = Collections.synchronizedSet(new HashSet<>());
+    private final EconomicSet<String> userRequestedSecurityProviders = EconomicSet.create();
 
     /**
      * A map of providers, identified by their names (see {@link Provider#getName()}), and the
@@ -68,7 +72,7 @@ public final class SecurityProvidersSupport {
      * structure is used instead of the (see javax.crypto.JceSecurity#verifyingProviders) map to
      * avoid keeping provider objects in the image heap.
      */
-    private final Map<String, Object> verifiedSecurityProviders = Collections.synchronizedMap(new HashMap<>());
+    private final EconomicMap<String, Object> verifiedSecurityProviders = ImageHeapMap.create("verifiedSecurityProviders");
 
     private Properties savedInitialSecurityProperties;
 
@@ -85,8 +89,8 @@ public final class SecurityProvidersSupport {
     }
 
     @Platforms(Platform.HOSTED_ONLY.class)
-    public void addVerifiedSecurityProvider(String key, Object value) {
-        verifiedSecurityProviders.put(key, value);
+    public void addVerifiedSecurityProvider(String key, Object verificationResult) {
+        verifiedSecurityProviders.put(key, verificationResult);
     }
 
     public Object getSecurityProviderVerificationResult(String key) {
@@ -113,7 +117,7 @@ public final class SecurityProvidersSupport {
      * qualified name (e.g., sun.security.provider.Sun), is either user-requested or reachable via a
      * security service.
      */
-    public boolean isSecurityProviderExpected(String providerName, String providerFQName) {
+    public boolean isSecurityProviderRequested(String providerName, String providerFQName) {
         return verifiedSecurityProviders.containsKey(providerName) || userRequestedSecurityProviders.contains(providerFQName);
     }
 
@@ -130,11 +134,55 @@ public final class SecurityProvidersSupport {
         }
     }
 
+    @Platforms(Platform.HOSTED_ONLY.class)
     public void setSavedInitialSecurityProperties(Properties savedSecurityProperties) {
         this.savedInitialSecurityProperties = savedSecurityProperties;
     }
 
     public Properties getSavedInitialSecurityProperties() {
         return savedInitialSecurityProperties;
+    }
+
+    public Provider loadBuiltInProvider(String provName, Debug debug) {
+        return switch (provName) {
+            case "SUN", "sun.security.provider.Sun" ->
+                isSecurityProviderRequested("SUN", "sun.security.provider.Sun") ? new sun.security.provider.Sun() : null;
+            case "SunRsaSign", "sun.security.rsa.SunRsaSign" ->
+                isSecurityProviderRequested("SunRsaSign", "sun.security.rsa.SunRsaSign") ? new sun.security.rsa.SunRsaSign() : null;
+            case "SunJCE", "com.sun.crypto.provider.SunJCE" ->
+                isSecurityProviderRequested("SunJCE", "com.sun.crypto.provider.SunJCE") ? new com.sun.crypto.provider.SunJCE() : null;
+            case "SunJSSE" ->
+                isSecurityProviderRequested("SunJSSE", "sun.security.ssl.SunJSSE") ? new sun.security.ssl.SunJSSE() : null;
+            case "SunEC" ->
+                isSecurityProviderRequested("SunEC", "sun.security.ec.SunEC") ? allocateSunECProvider() : null;
+            case "Apple", "apple.security.AppleProvider" -> {
+                try {
+                    Class<?> c = Class.forName("apple.security.AppleProvider");
+                    if (Provider.class.isAssignableFrom(c)) {
+                        yield (Provider) c.getDeclaredConstructor().newInstance();
+                    }
+                } catch (Exception ex) {
+                    if (debug != null) {
+                        debug.println("Error loading provider Apple");
+                        ex.printStackTrace();
+                    }
+                }
+                yield null;
+            }
+            default -> null;
+        };
+    }
+
+    public static boolean isBuiltInProvider(String provName) {
+        return switch (provName) {
+            case "SUN", "sun.security.provider.Sun",
+                            "SunRsaSign", "sun.security.rsa.SunRsaSign",
+                            "SunJCE", "com.sun.crypto.provider.SunJCE",
+                            "SunJSSE",
+                            "SunEC",
+                            "Apple", "apple.security.AppleProvider" ->
+                true;
+            default -> false;
+        };
     }
 }